requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
python-decouple redis
asyncpg-lite